+++ e2e/mv_test.go
@@ -10,6 +10,7 @@ import (
 	"gotest.tools/v3/icmd"
 )
 
+// mv s3://bucket/key dir/
 func TestMoveSingleS3ObjectToLocal(t *testing.T) {
 	t.Parallel()
 
@@ -46,7 +47,8 @@ func TestMoveSingleS3ObjectToLocal(t *testing.T) {
 	assertError(t, err, errS3NoSuchKey)
 }
 
-func TestMoveMultipleFlatS3ObjectsToLocal(t *testing.T) {
+// mv s3://bucket/key dir/
+func TestMoveMultipleS3ObjectsToLocal(t *testing.T) {
 	t.Parallel()
 
 	bucket := s3BucketFromTestName(t)
@@ -96,6 +98,7 @@ func TestMoveMultipleFlatS3ObjectsToLocal(t *testing.T) {
 	}
 }
 
+// mv file s3://bucket
 func TestMoveSingleFileToS3(t *testing.T) {
 	t.Parallel()
 
@@ -133,6 +136,7 @@ func TestMoveSingleFileToS3(t *testing.T) {
 	assert.Assert(t, ensureS3Object(s3client, bucket, filename, content))
 }
 
+// mv dir/* s3://bucket
 func TestMoveMultipleFilesToS3(t *testing.T) {
 	t.Parallel()
 
@@ -184,6 +188,7 @@ func TestMoveMultipleFilesToS3(t *testing.T) {
 	}
 }
 
+// mv s3://bucket/object s3://bucket2/object
 func TestMoveSingleS3ObjectToS3(t *testing.T) {
 	t.Parallel()
 
@@ -222,6 +227,7 @@ func TestMoveSingleS3ObjectToS3(t *testing.T) {
 	assert.Assert(t, ensureS3Object(s3client, bucket, "dst/"+filename, content))
 }
 
+// mv s3://bucket/object s3://bucket2/object
 func TestMoveSingleS3ObjectIntoAnotherBucket(t *testing.T) {
 	t.Parallel()
 
@@ -262,6 +268,7 @@ func TestMoveSingleS3ObjectIntoAnotherBucket(t *testing.T) {
 	assert.Assert(t, ensureS3Object(s3client, dstbucket, filename, content))
 }
 
+// mv s3://bucket/* s3://bucket2/prefix/
 func TestMoveMultipleS3ObjectsToS3(t *testing.T) {
 	t.Parallel()