+++ CHANGELOG.md
@@ -31,6 +31,8 @@ This is a major release with many breaking changes.
   concurrency and part size flags are now part of the `cp/mv` command. New
   replacement flags are `--concurrency | -c` and `--part-size | -p`. ([#110](https://github.com/peak/s5cmd/pull/110))
 - s5cmd `cp` command follows symbolic links by default (only when uploading to s3 from local filesystem). Use `--no-follow-symlinks` flag to disable this feature.
+- Dropped `-parents` flag from copy command. Copy behaviour has changed to preserve the directory hierarchy as a default.
+  Optional `-flatten` flag is added to flatten directory structure.
 
 #### Features